Added Android debug mode to debug ignored events.
Been providing this app to users to help identify the root
cause of issues with their gamepad, currently attempting to
debug an issue with a particular gamepad, the XBox Elite 2,
that does not appear to work the same as a normal XBox
controller in that the Share/select button is not working.
Hopefully this will help identify what event that particular
controller is generating when the Share/select button is
pressed.